cleanup echo effect
